feat: production deployment setup + security hardening
## Deployment Infrastructure
- Multi-stage Dockerfile (deps + runner stages) with curl healthcheck
- docker-compose.yml (development) with PostgreSQL + hotreload
- docker-compose.prod.yml with isolated network, resource limits, logging
- ecosystem.config.js for PM2 cluster mode
- .env.example with all Nanei-specific variables
- Updated .dockerignore with comprehensive exclusions

## Deployment Automation
- GitHub Actions workflow (.github/workflows/deploy.yml) for auto-deploy on main push
- deploy/deploy.sh: production deployment script with migrations + healthcheck
- deploy/nginx.conf: reverse proxy configuration with SSL/TLS, gzip, rate-limiting
- deploy/initdb/01-init.sql: PostgreSQL initialization (extensions)
- deploy/README.md: comprehensive deployment guide
- deploy/DEPLOYMENT_CHECKLIST.md: 50+ point pre-launch checklist
- DEPLOYMENT_SUMMARY.md: overview of all deployment files

## Security Hardening (Phase 1)
- Fixed IDOR: removed :id URL parameters, using JWT req.user.id
- Added webhook signature validation: HMAC verification for Wave + Orange Money
- Implemented OTP rate-limiting: 3 attempts → 15min lockout
- Token encryption: OTP codes hashed with bcrypt (no plaintext)
- Payment amount validation: 100 XOF min, 10M XOF max, 2 decimals
- RGPD-compliant account deletion: hard delete (not anonymization)
- Container healthcheck: /health endpoint with curl
- Removed duplicate dependencies: bcryptjs removed, keeping bcrypt only
- DECIMAL(10,2) for monetary columns: prevents float rounding errors

## Middleware & Services
- webhookSignature.middleware.js: HMAC validation for payments
- paymentValidation.middleware.js: amount range + decimal validation
- rgpd.service.js: complete hard-delete with cascade
- Updated UserOtp model: failedAttempts + lockedUntil fields

## Configuration Updates
- app.js: added /health route + rawBody capture for webhook signatures
- auth.service.js: OTP generation with bcrypt hash
- account.service.js: OTP verification with rate-limiting
- auth.route.js: IDOR fix, removed :id parameters
- paiement.controller.js: webhook signature validation
- paiement.route.js: payment amount validation middleware

All files production-ready, compatible with Sign deployment infrastructure.

+++ src/models/paiement.model.js
@@ -20,11 +20,11 @@ const Paiement = sequelize.define('Paiement', {
     references: { model: 'utilisateur', key: 'id' },
   },
   prixTotal: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(10, 2), // SÉCURITÉ: DECIMAL pour argent (pas FLOAT qui arrondit mal)
     allowNull: false,
   },
   montantPaye: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(10, 2), // SÉCURITÉ: DECIMAL pour argent (pas FLOAT qui arrondit mal)
     defaultValue: 0,
     allowNull: false,
   },
